+++ src/library/Client/ServiceClients/Handlers/ServiceClientCredentialsHttpHandler.cs
@@ -3,11 +3,11 @@
 // Licensed under the MIT license.
 // --------------------------------------------------------------------------------------------
 
+using k8s.Autorest;
 using System;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.Rest;
 
 namespace Microsoft.BridgeToKubernetes.Library.ServiceClients.Handlers
 {